// src/types/invoice.ts
export type Invoice = {
    id: number;
    client_id: number;
    seller_user_id: number | null;
    order_id: number | null;
    created_by: number;
    invoice_no: string;
    invoice_date: string;
    challan_no: string | null;
    challan_date: string | null;
    eway_bill_no: string | null;
    transport_name: string | null;
    transport_id: string | null;
    place_of_supply: string | null;
    currency: string;
    seller_name: string;
    seller_tagline: string | null;
    seller_address: string | null;
    seller_phone: string | null;
    seller_email: string | null;
    seller_website: string | null;
    seller_pan: string | null;
    seller_gstin: string | null;
    bank_name: string | null;
    bank_branch: string | null;
    bank_account_no: string | null;
    bank_ifsc: string | null;
    bank_upi_id: string | null;
    upi_qr_image_url: string | null;
    subtotal: number;
    tax_total: number;
    discount_total: number;
    round_off: number;
    total_amount: number;
    amount_in_words: string | null;
    notes: string | null;
    terms_conditions: string | null;
    declaration: string | null;
    customer_signature_label: string | null;
    authorized_signatory_label: string | null;
    footer_note: string | null;
    created_at: Date;
    updated_at: Date;
}